Move ParamId to burn-std to replace DistributedParamId (type alias)
