+++ crates/burn-train/src/learner/sharder.rs
@@ -1,7 +1,6 @@
 use burn_core::{
     Tensor,
     module::{Module, ModuleMapper, Param},
-    tensor::backend::distributed::DistributedParamId,
 };
 
 use crate::{Learner, LearningComponentsTypes};
@@ -12,7 +11,7 @@ pub struct ModuleSharder;
 impl ModuleMapper for ModuleSharder {
     fn map_float<const D: usize>(&mut self, param: Param<Tensor<D>>) -> Param<Tensor<D>> {
         let (id, tensor, mapper) = param.consume();
-        let tensor = tensor.set_distributed(DistributedParamId::from(id.val()));
+        let tensor = tensor.set_distributed(id);
         Param::from_mapped_value(id, tensor, mapper)
     }
 }
